Dynamically create nav-graph and remove child classes
The navigation graph is now created based on an array containing
all the map data instead of defining it hard-coded in an xml file.
At the same time, this array of map data is now used to populate
the maps, removing the need for all the child classes of all the
maps, previously inheriting from the 'BaseFragment'. This fragment
is now known as the 'MapFragment'.